refactor: 사용되지 않는 Dao 인터페이스 제거 및 Dao 클래스명 변경
