+++ src/main/java/wooteco/subway/dao/StationDao.java
@@ -1,4 +1,4 @@
-package wooteco.subway.dao.station;
+package wooteco.subway.dao;
 
 import java.util.HashMap;
 import java.util.List;
@@ -8,18 +8,16 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import wooteco.subway.domain.Id;
-import wooteco.subway.domain.Name;
 import wooteco.subway.domain.Station;
 
 @Repository
-public class StationDaoH2 implements StationDao {
+public class StationDao {
 
     private final JdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
     private final RowMapper<Station> rowMapper;
 
-    public StationDaoH2(final JdbcTemplate jdbcTemplate, final DataSource source) {
+    public StationDao(final JdbcTemplate jdbcTemplate, final DataSource source) {
         this.jdbcTemplate = jdbcTemplate;
         this.jdbcInsert = new SimpleJdbcInsert(source)
             .withTableName("STATION")
@@ -31,7 +29,6 @@ public class StationDaoH2 implements StationDao {
         };
     }
 
-    @Override
     public Station save(final Station station) {
         Map<String, String> params = new HashMap<>();
         params.put("name", station.getName());
@@ -41,13 +38,11 @@ public class StationDaoH2 implements StationDao {
         return new Station(id, station.getName());
     }
 
-    @Override
     public List<Station> showAll() {
         String statement = "SELECT * FROM STATION";
         return jdbcTemplate.query(statement, rowMapper);
     }
 
-    @Override
     public int delete(final long id) {
         String statement = "DELETE FROM station WHERE id = ?";
         return jdbcTemplate.update(statement, id);
