--- src/main/java/wooteco/subway/dao/station/StationDao.java
@@ -1,13 +1,0 @@
-package wooteco.subway.dao.station;
-
-import java.util.List;
-import wooteco.subway.domain.Station;
-
-public interface StationDao {
-
-    Station save(Station station);
-
-    List<Station> showAll();
-
-    int delete(long id);
-}
