+++ src/main/java/wooteco/subway/service/LineService.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import javax.validation.Valid;
 import org.springframework.stereotype.Service;
-import wooteco.subway.dao.line.LineDao;
+import wooteco.subway.dao.LineDao;
 import wooteco.subway.domain.Line;
 import wooteco.subway.exception.line.NotFoundLineException;
 import wooteco.subway.service.dto.LineServiceDto;
