+++ src/main/java/wooteco/subway/service/StationService.java
@@ -3,9 +3,8 @@ package wooteco.subway.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
-import wooteco.subway.domain.Name;
+import wooteco.subway.dao.StationDao;
 import wooteco.subway.exception.station.NotFoundStationException;
-import wooteco.subway.dao.station.StationDao;
 import wooteco.subway.service.dto.StationServiceDto;
 import wooteco.subway.domain.Station;
